refactor: rename wrappers append _wrapper
