+++ src/wrappers/contract_wrappers/e_r_c721_collateralized_simple_interest_terms_contract_wrapper.ts
